requirements.txt
numpy==1.26.4
pandas==2.2.3

fastapi==0.115.6
uvicorn==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1

sentence-transformers==3.0.1
scikit-learn==1.5.2
joblib==1.4.2


orjson==3.10.7
xxhash==3.5.0

pytest==8.3.3
